completed unit-tests in ags_thread_file_xml_test.c

+++ ags/thread/ags_timestamp.h
@@ -54,7 +54,7 @@ struct _AgsTimestamp
     }ags_offset;
   }timer;
 
-  guint delay;
+  gdouble delay;
   guint attack;
 };
